Switch trees and shrubs, use custom strings
In the map editor, the "trees" (2nd) button places what seem like
shrubs (which change as you expand out), and the "scrub" (5th) button
places what seem like trees. This may be intended but is unintuitive.
The commit renames the 2nd button to "shrubs" and the 5th to "trees"
and handles the vars/functions accordingly.
Button names use custom strings instead of the .eng file.