Allow setting trade route cost for city
Add a button for setting the cost of opening a trade route with the
selected city. Adjust buffer sizes to allow for 5 digit numbers.

Improve editor empire map deselection handling - going back while
having selected an object deselects it first; same behaviour as on
the empire map when playing a map.